+++ src/main/java/com/eaaslan/cinema_booking_system/model/Theatre.java
@@ -1,0 +1,32 @@
+package com.eaaslan.cinema_booking_system.model;
+
+import jakarta.persistence.*;
+import lombok.*;
+import java.util.List;
+
+@Entity
+@Table(name = "theatres")
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class Theatre {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(name = "name", nullable = false)
+    private String name;
+
+    @OneToMany(mappedBy = "theatre", cascade = CascadeType.ALL)
+    private List<Room> screenings;
+
+    @Override
+    public String toString() {
+        return "Theatre{" +
+                "id=" + id +
+                ", name='" + name + '\'' +
+                '}';
+    }
+}