+++ CHANGELOG.md
@@ -1,7 +1,14 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## [0.1.0] - 2023-07-20
+## [0.2.0] - 2024-07-19
+### Added
+-Core entities (Movie, Theatre, Room, Seat, Screening, Booking)
+-JPA annotations for database mapping
+-Lombok integration for reducing boilerplate code
+
+## [0.1.0] - 2024-07-19
 ### Added
 - Initial project setup using Spring Initializr
 - Basic README and CHANGELOG
+